refactored file naming convention. The Tabs component is now seperate from Timeline
