--- src/components/Activities/index.ts
@@ -1,2 +1,0 @@
-export * from './Activities'
-export * from './AreaSurface'
