+++ src/components/Timeline/index.ts
@@ -1,0 +1,2 @@
+export * from './Timeline'
+export * from './IntervalSurface'
